Modify: 아이콘 경로 수정
Add: 검색기능 최적화

Add: Vector 아이콘 추가 & camera, clip 아이콘 추가
