--- src/assets/Anchor.js
@@ -1,37 +1,0 @@
-import React from 'react';
-
-const Anchor = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12.0001 4.66602C11.1716 4.66602 10.5001 5.33759 10.5001 6.16602C10.5001 6.99444 11.1716 7.66602 12.0001 7.66602C12.8285 7.66602 13.5001 6.99444 13.5001 6.16602C13.5001 5.33759 12.8285 4.66602 12.0001 4.66602ZM8.50006 6.16602C8.50006 4.23302 10.0671 2.66602 12.0001 2.66602C13.9331 2.66602 15.5001 4.23302 15.5001 6.16602C15.5001 8.09901 13.9331 9.66602 12.0001 9.66602C10.0671 9.66602 8.50006 8.09901 8.50006 6.16602Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12.0001 7.66602C12.5523 7.66602 13.0001 8.11376 13.0001 8.66605V20.3327C13.0001 20.885 12.5523 21.3327 12.0001 21.3327C11.4478 21.3327 11.0001 20.885 11.0001 20.3327V8.66605C11.0001 8.11376 11.4478 7.66602 12.0001 7.66602Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M2.95964 11.2923C3.14718 11.1047 3.40153 10.9994 3.66675 10.9994H6.16675C6.71903 10.9994 7.16675 11.4471 7.16675 11.9994C7.16675 12.5516 6.71903 12.9994 6.16675 12.9994H4.73522C4.95207 14.5749 5.67725 16.0474 6.81463 17.1848C8.1899 18.5601 10.0552 19.3327 12.0001 19.3327C13.945 19.3327 15.8103 18.5601 17.1855 17.1848C18.3229 16.0474 19.0481 14.5749 19.2649 12.9994H17.8334C17.2811 12.9994 16.8334 12.5516 16.8334 11.9994C16.8334 11.4471 17.2811 10.9994 17.8334 10.9994H20.3334C20.8857 10.9994 21.3334 11.4471 21.3334 11.9994C21.3334 14.4747 20.3501 16.8487 18.5997 18.599C16.8494 20.3494 14.4754 21.3327 12.0001 21.3327C9.52471 21.3327 7.15076 20.3494 5.40042 18.599C3.65008 16.8487 2.66675 14.4747 2.66675 11.9994C2.66675 11.7341 2.77211 11.4798 2.95964 11.2923Z"
-          fill="#999999"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Anchor;
