--- src/assets/Arrow.js
@@ -1,31 +1,0 @@
-import React from 'react';
-
-const Arrow = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M15.1666 7C15.1666 6.44772 15.6143 6 16.1666 6H21.1666C21.7188 6 22.1666 6.44772 22.1666 7V12C22.1666 12.5523 21.7188 13 21.1666 13C20.6143 13 20.1666 12.5523 20.1666 12V8H16.1666C15.6143 8 15.1666 7.55228 15.1666 7Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M21.8737 6.29289C22.2642 6.68342 22.2642 7.31658 21.8737 7.70711L13.957 15.6238C13.5665 16.0143 12.9333 16.0143 12.5428 15.6238L9.08325 12.1642L3.54036 17.7071C3.14983 18.0976 2.51667 18.0976 2.12615 17.7071C1.73562 17.3166 1.73562 16.6834 2.12615 16.2929L8.37615 10.0429C8.76667 9.65237 9.39983 9.65237 9.79036 10.0429L13.2499 13.5025L20.4595 6.29289C20.85 5.90237 21.4832 5.90237 21.8737 6.29289Z"
-          fill="#999999"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Arrow;
