--- src/assets/Chat.js
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const Anchor = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M6.16667 5.5C5.98986 5.5 5.82029 5.57024 5.69526 5.69526C5.57024 5.82029 5.5 5.98986 5.5 6.16667V17.0858L7.12623 15.4596C7.31376 15.272 7.56812 15.1667 7.83333 15.1667H17.8333C18.0101 15.1667 18.1797 15.0964 18.3047 14.9714C18.4298 14.8464 18.5 14.6768 18.5 14.5V6.16667C18.5 5.98986 18.4298 5.82029 18.3047 5.69526C18.1797 5.57024 18.0101 5.5 17.8333 5.5H6.16667ZM4.28105 4.28105C4.78115 3.78095 5.45942 3.5 6.16667 3.5H17.8333C18.5406 3.5 19.2189 3.78095 19.719 4.28105C20.219 4.78115 20.5 5.45942 20.5 6.16667V14.5C20.5 15.2072 20.219 15.8855 19.719 16.3856C19.2189 16.8857 18.5406 17.1667 17.8333 17.1667H8.24755L5.20711 20.2071C4.92111 20.4931 4.49099 20.5787 4.11732 20.4239C3.74364 20.2691 3.5 19.9045 3.5 19.5V6.16667C3.5 5.45942 3.78095 4.78115 4.28105 4.28105Z"
-          fill="white"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Anchor;
