--- src/assets/Cube.js
@@ -1,37 +1,0 @@
-import React from 'react';
-
-const Cube = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M10.6688 3.02399C11.0737 2.79075 11.5327 2.66797 12 2.66797C12.4672 2.66797 12.9263 2.79074 13.3311 3.02395L13.3333 3.02523L19.1667 6.35856C19.5717 6.59238 19.908 6.92859 20.142 7.33345C20.3761 7.73831 20.4995 8.1976 20.5 8.66523V15.334C20.4995 15.8016 20.3761 16.2609 20.142 16.6657C19.908 17.0706 19.5716 17.4068 19.1667 17.6406L19.1628 17.6428L13.3333 20.974L13.3314 20.9751C12.9265 21.2084 12.4674 21.3312 12 21.3312C11.5326 21.3312 11.0735 21.2084 10.6685 20.975L10.6667 20.974L4.83719 17.6428L4.83333 17.6406C4.42835 17.4068 4.09197 17.0706 3.85795 16.6657C3.62392 16.2609 3.50048 15.8016 3.5 15.334V8.66523C3.50048 8.1976 3.62392 7.73831 3.85795 7.33345C4.09197 6.92859 4.42835 6.59238 4.83333 6.35857L4.83719 6.35634L10.6688 3.02399ZM12 4.66797C11.883 4.66797 11.768 4.69877 11.6667 4.75729L11.6628 4.75951L5.83185 8.09148C5.73127 8.14989 5.64771 8.23362 5.58949 8.33434C5.53105 8.43544 5.50019 8.55011 5.5 8.66688V15.3323C5.50019 15.4491 5.53105 15.5637 5.58949 15.6648C5.64769 15.7655 5.73121 15.8492 5.83175 15.9077L5.83333 15.9086L11.6667 19.2419C11.768 19.3004 11.883 19.3312 12 19.3312C12.117 19.3312 12.232 19.3004 12.3333 19.2419L12.3372 19.2397L18.1667 15.9086L18.1682 15.9077C18.2688 15.8492 18.3523 15.7655 18.4105 15.6648C18.469 15.5636 18.4999 15.4488 18.5 15.3319V8.66729C18.4999 8.55038 18.469 8.43555 18.4105 8.33434C18.3523 8.23362 18.2687 8.14989 18.1681 8.09148L18.1667 8.09062L12.3333 4.75729C12.232 4.69878 12.117 4.66797 12 4.66797Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12 10.9996C12.5523 10.9996 13 11.4473 13 11.9996V20.3996C13 20.9519 12.5523 21.3996 12 21.3996C11.4477 21.3996 11 20.9519 11 20.3996V11.9996C11 11.4473 11.4477 10.9996 12 10.9996Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M3.85941 7.29887C4.13595 6.82081 4.74768 6.65744 5.22574 6.93399L12 10.8527L18.7743 6.93399C19.2524 6.65744 19.8641 6.82081 20.1406 7.29887C20.4172 7.77693 20.2538 8.38866 19.7757 8.6652L12.5007 12.8735C12.191 13.0527 11.8091 13.0527 11.4993 12.8735L4.2243 8.6652C3.74623 8.38866 3.58287 7.77693 3.85941 7.29887Z"
-          fill="#999999"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Cube;
