--- src/assets/Folder.js
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const Folder = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M5.33341 5.5C5.1566 5.5 4.98703 5.57024 4.86201 5.69526C4.73699 5.82029 4.66675 5.98986 4.66675 6.16667V17.8333C4.66675 18.0101 4.73699 18.1797 4.86201 18.3047C4.98703 18.4298 5.1566 18.5 5.33341 18.5H18.6667C18.8436 18.5 19.0131 18.4298 19.1382 18.3047C19.2632 18.1797 19.3334 18.0101 19.3334 17.8333V8.66667C19.3334 8.48986 19.2632 8.32029 19.1382 8.19526C19.0131 8.07024 18.8436 8 18.6667 8H11.1667C10.8324 8 10.5202 7.8329 10.3347 7.5547L8.9649 5.5H5.33341ZM3.4478 4.28105C3.94789 3.78095 4.62617 3.5 5.33341 3.5H9.50008C9.83443 3.5 10.1467 3.6671 10.3321 3.9453L11.7019 6H18.6667C19.374 6 20.0523 6.28095 20.5524 6.78105C21.0525 7.28115 21.3334 7.95942 21.3334 8.66667V17.8333C21.3334 18.5406 21.0525 19.2189 20.5524 19.719C20.0523 20.219 19.374 20.5 18.6667 20.5H5.33341C4.62617 20.5 3.94789 20.219 3.4478 19.719C2.9477 19.2189 2.66675 18.5406 2.66675 17.8333V6.16667C2.66675 5.45942 2.9477 4.78115 3.4478 4.28105Z"
-          fill="#999999"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Folder;
