+++ src/assets/Icons.js
@@ -32,7 +32,6 @@ export const Info = () => {
     </div>
   );
 };
-
 export const Vector = () => {
   return (
     <div className="list__icon vector">
@@ -97,6 +96,572 @@ export const Success = () => {
           clipRule="evenodd"
           d="M22 12C22 17.5228 17.5228 22 12 22C6.47715 22 2 17.5228 2 12C2 6.47715 6.47715 2 12 2C17.5228 2 22 6.47715 22 12ZM9.3824 11.0689C9.50441 11.1213 9.61475 11.1975 9.707 11.293L11 12.586L14.293 9.29302C14.3852 9.19751 14.4956 9.12133 14.6176 9.06892C14.7396 9.01651 14.8708 8.98892 15.0036 8.98777C15.1364 8.98662 15.2681 9.01192 15.391 9.0622C15.5138 9.11248 15.6255 9.18673 15.7194 9.28063C15.8133 9.37452 15.8875 9.48617 15.9378 9.60907C15.9881 9.73196 16.0134 9.86364 16.0122 9.99642C16.0111 10.1292 15.9835 10.2604 15.9311 10.3824C15.8787 10.5044 15.8025 10.6148 15.707 10.707L11.707 14.707C11.5195 14.8945 11.2652 14.9998 11 14.9998C10.7348 14.9998 10.4805 14.8945 10.293 14.707L8.293 12.707C8.19749 12.6148 8.1213 12.5044 8.0689 12.3824C8.01649 12.2604 7.9889 12.1292 7.98775 11.9964C7.98659 11.8636 8.0119 11.732 8.06218 11.6091C8.11246 11.4862 8.18671 11.3745 8.2806 11.2806C8.3745 11.1867 8.48615 11.1125 8.60904 11.0622C8.73194 11.0119 8.86362 10.9866 8.9964 10.9878C9.12918 10.9889 9.2604 11.0165 9.3824 11.0689Z"
           fill="white"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const AddChannel = ({
+  setCreateType,
+  setIsCreating,
+  setIsEditing,
+  type,
+}) => (
+  <svg
+    width="14"
+    height="14"
+    viewBox="0 0 14 14"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    onClick={() => {
+      setCreateType(type);
+      setIsCreating(prevState => !prevState);
+      setIsEditing(false);
+    }}
+  >
+    <path
+      d="M7 0C3.13438 0 0 3.13438 0 7C0 10.8656 3.13438 14 7 14C10.8656 14 14 10.8656 14 7C14 3.13438 10.8656 0 7 0ZM11 7.5H7.5V11H6.5V7.5H3V6.5H6.5V3H7.5V6.5H11V7.5Z"
+      fill="white"
+      fillOpacity="0.66"
+    />
+  </svg>
+);
+
+export const Anchor = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12.0001 4.66602C11.1716 4.66602 10.5001 5.33759 10.5001 6.16602C10.5001 6.99444 11.1716 7.66602 12.0001 7.66602C12.8285 7.66602 13.5001 6.99444 13.5001 6.16602C13.5001 5.33759 12.8285 4.66602 12.0001 4.66602ZM8.50006 6.16602C8.50006 4.23302 10.0671 2.66602 12.0001 2.66602C13.9331 2.66602 15.5001 4.23302 15.5001 6.16602C15.5001 8.09901 13.9331 9.66602 12.0001 9.66602C10.0671 9.66602 8.50006 8.09901 8.50006 6.16602Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12.0001 7.66602C12.5523 7.66602 13.0001 8.11376 13.0001 8.66605V20.3327C13.0001 20.885 12.5523 21.3327 12.0001 21.3327C11.4478 21.3327 11.0001 20.885 11.0001 20.3327V8.66605C11.0001 8.11376 11.4478 7.66602 12.0001 7.66602Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M2.95964 11.2923C3.14718 11.1047 3.40153 10.9994 3.66675 10.9994H6.16675C6.71903 10.9994 7.16675 11.4471 7.16675 11.9994C7.16675 12.5516 6.71903 12.9994 6.16675 12.9994H4.73522C4.95207 14.5749 5.67725 16.0474 6.81463 17.1848C8.1899 18.5601 10.0552 19.3327 12.0001 19.3327C13.945 19.3327 15.8103 18.5601 17.1855 17.1848C18.3229 16.0474 19.0481 14.5749 19.2649 12.9994H17.8334C17.2811 12.9994 16.8334 12.5516 16.8334 11.9994C16.8334 11.4471 17.2811 10.9994 17.8334 10.9994H20.3334C20.8857 10.9994 21.3334 11.4471 21.3334 11.9994C21.3334 14.4747 20.3501 16.8487 18.5997 18.599C16.8494 20.3494 14.4754 21.3327 12.0001 21.3327C9.52471 21.3327 7.15076 20.3494 5.40042 18.599C3.65008 16.8487 2.66675 14.4747 2.66675 11.9994C2.66675 11.7341 2.77211 11.4798 2.95964 11.2923Z"
+          fill="#999999"
+        />
+      </svg>
+    </div>
+  );
+};
+
+const Chat = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M6.16667 5.5C5.98986 5.5 5.82029 5.57024 5.69526 5.69526C5.57024 5.82029 5.5 5.98986 5.5 6.16667V17.0858L7.12623 15.4596C7.31376 15.272 7.56812 15.1667 7.83333 15.1667H17.8333C18.0101 15.1667 18.1797 15.0964 18.3047 14.9714C18.4298 14.8464 18.5 14.6768 18.5 14.5V6.16667C18.5 5.98986 18.4298 5.82029 18.3047 5.69526C18.1797 5.57024 18.0101 5.5 17.8333 5.5H6.16667ZM4.28105 4.28105C4.78115 3.78095 5.45942 3.5 6.16667 3.5H17.8333C18.5406 3.5 19.2189 3.78095 19.719 4.28105C20.219 4.78115 20.5 5.45942 20.5 6.16667V14.5C20.5 15.2072 20.219 15.8855 19.719 16.3856C19.2189 16.8857 18.5406 17.1667 17.8333 17.1667H8.24755L5.20711 20.2071C4.92111 20.4931 4.49099 20.5787 4.11732 20.4239C3.74364 20.2691 3.5 19.9045 3.5 19.5V6.16667C3.5 5.45942 3.78095 4.78115 4.28105 4.28105Z"
+          fill="white"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export default Chat;
+
+export const Arrow = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M15.1666 7C15.1666 6.44772 15.6143 6 16.1666 6H21.1666C21.7188 6 22.1666 6.44772 22.1666 7V12C22.1666 12.5523 21.7188 13 21.1666 13C20.6143 13 20.1666 12.5523 20.1666 12V8H16.1666C15.6143 8 15.1666 7.55228 15.1666 7Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M21.8737 6.29289C22.2642 6.68342 22.2642 7.31658 21.8737 7.70711L13.957 15.6238C13.5665 16.0143 12.9333 16.0143 12.5428 15.6238L9.08325 12.1642L3.54036 17.7071C3.14983 18.0976 2.51667 18.0976 2.12615 17.7071C1.73562 17.3166 1.73562 16.6834 2.12615 16.2929L8.37615 10.0429C8.76667 9.65237 9.39983 9.65237 9.79036 10.0429L13.2499 13.5025L20.4595 6.29289C20.85 5.90237 21.4832 5.90237 21.8737 6.29289Z"
+          fill="#999999"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const Camera = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M8.66787 3.9453C8.85333 3.6671 9.16557 3.5 9.49992 3.5H14.4999C14.8343 3.5 15.1465 3.6671 15.332 3.9453L16.7018 6H19.4999C20.2072 6 20.8854 6.28095 21.3855 6.78105C21.8856 7.28115 22.1666 7.95942 22.1666 8.66667V17.8333C22.1666 18.5406 21.8856 19.2189 21.3855 19.719C20.8854 20.219 20.2072 20.5 19.4999 20.5H4.49992C3.79267 20.5 3.1144 20.219 2.6143 19.719C2.1142 19.2189 1.83325 18.5406 1.83325 17.8333V8.66667C1.83325 7.95942 2.1142 7.28115 2.6143 6.78105C3.1144 6.28095 3.79267 6 4.49992 6H7.29807L8.66787 3.9453ZM10.0351 5.5L8.6653 7.5547C8.47984 7.8329 8.16761 8 7.83325 8H4.49992C4.32311 8 4.15354 8.07024 4.02851 8.19526C3.90349 8.32029 3.83325 8.48986 3.83325 8.66667V17.8333C3.83325 18.0101 3.90349 18.1797 4.02851 18.3047C4.15354 18.4298 4.32311 18.5 4.49992 18.5H19.4999C19.6767 18.5 19.8463 18.4298 19.9713 18.3047C20.0963 18.1797 20.1666 18.0101 20.1666 17.8333V8.66667C20.1666 8.48986 20.0963 8.32029 19.9713 8.19526C19.8463 8.07024 19.6767 8 19.4999 8H16.1666C15.8322 8 15.52 7.8329 15.3345 7.5547L13.9647 5.5H10.0351Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12.0001 10.5C10.7114 10.5 9.66675 11.5447 9.66675 12.8333C9.66675 14.122 10.7114 15.1667 12.0001 15.1667C13.2887 15.1667 14.3334 14.122 14.3334 12.8333C14.3334 11.5447 13.2887 10.5 12.0001 10.5ZM7.66675 12.8333C7.66675 10.4401 9.60685 8.5 12.0001 8.5C14.3933 8.5 16.3334 10.4401 16.3334 12.8333C16.3334 15.2266 14.3933 17.1667 12.0001 17.1667C9.60685 17.1667 7.66675 15.2266 7.66675 12.8333Z"
+          fill="#1B1B1B"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const Cube = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M10.6688 3.02399C11.0737 2.79075 11.5327 2.66797 12 2.66797C12.4672 2.66797 12.9263 2.79074 13.3311 3.02395L13.3333 3.02523L19.1667 6.35856C19.5717 6.59238 19.908 6.92859 20.142 7.33345C20.3761 7.73831 20.4995 8.1976 20.5 8.66523V15.334C20.4995 15.8016 20.3761 16.2609 20.142 16.6657C19.908 17.0706 19.5716 17.4068 19.1667 17.6406L19.1628 17.6428L13.3333 20.974L13.3314 20.9751C12.9265 21.2084 12.4674 21.3312 12 21.3312C11.5326 21.3312 11.0735 21.2084 10.6685 20.975L10.6667 20.974L4.83719 17.6428L4.83333 17.6406C4.42835 17.4068 4.09197 17.0706 3.85795 16.6657C3.62392 16.2609 3.50048 15.8016 3.5 15.334V8.66523C3.50048 8.1976 3.62392 7.73831 3.85795 7.33345C4.09197 6.92859 4.42835 6.59238 4.83333 6.35857L4.83719 6.35634L10.6688 3.02399ZM12 4.66797C11.883 4.66797 11.768 4.69877 11.6667 4.75729L11.6628 4.75951L5.83185 8.09148C5.73127 8.14989 5.64771 8.23362 5.58949 8.33434C5.53105 8.43544 5.50019 8.55011 5.5 8.66688V15.3323C5.50019 15.4491 5.53105 15.5637 5.58949 15.6648C5.64769 15.7655 5.73121 15.8492 5.83175 15.9077L5.83333 15.9086L11.6667 19.2419C11.768 19.3004 11.883 19.3312 12 19.3312C12.117 19.3312 12.232 19.3004 12.3333 19.2419L12.3372 19.2397L18.1667 15.9086L18.1682 15.9077C18.2688 15.8492 18.3523 15.7655 18.4105 15.6648C18.469 15.5636 18.4999 15.4488 18.5 15.3319V8.66729C18.4999 8.55038 18.469 8.43555 18.4105 8.33434C18.3523 8.23362 18.2687 8.14989 18.1681 8.09148L18.1667 8.09062L12.3333 4.75729C12.232 4.69878 12.117 4.66797 12 4.66797Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12 10.9996C12.5523 10.9996 13 11.4473 13 11.9996V20.3996C13 20.9519 12.5523 21.3996 12 21.3996C11.4477 21.3996 11 20.9519 11 20.3996V11.9996C11 11.4473 11.4477 10.9996 12 10.9996Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M3.85941 7.29887C4.13595 6.82081 4.74768 6.65744 5.22574 6.93399L12 10.8527L18.7743 6.93399C19.2524 6.65744 19.8641 6.82081 20.1406 7.29887C20.4172 7.77693 20.2538 8.38866 19.7757 8.6652L12.5007 12.8735C12.191 13.0527 11.8091 13.0527 11.4993 12.8735L4.2243 8.6652C3.74623 8.38866 3.58287 7.77693 3.85941 7.29887Z"
+          fill="#999999"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const Folder = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M5.33341 5.5C5.1566 5.5 4.98703 5.57024 4.86201 5.69526C4.73699 5.82029 4.66675 5.98986 4.66675 6.16667V17.8333C4.66675 18.0101 4.73699 18.1797 4.86201 18.3047C4.98703 18.4298 5.1566 18.5 5.33341 18.5H18.6667C18.8436 18.5 19.0131 18.4298 19.1382 18.3047C19.2632 18.1797 19.3334 18.0101 19.3334 17.8333V8.66667C19.3334 8.48986 19.2632 8.32029 19.1382 8.19526C19.0131 8.07024 18.8436 8 18.6667 8H11.1667C10.8324 8 10.5202 7.8329 10.3347 7.5547L8.9649 5.5H5.33341ZM3.4478 4.28105C3.94789 3.78095 4.62617 3.5 5.33341 3.5H9.50008C9.83443 3.5 10.1467 3.6671 10.3321 3.9453L11.7019 6H18.6667C19.374 6 20.0523 6.28095 20.5524 6.78105C21.0525 7.28115 21.3334 7.95942 21.3334 8.66667V17.8333C21.3334 18.5406 21.0525 19.2189 20.5524 19.719C20.0523 20.219 19.374 20.5 18.6667 20.5H5.33341C4.62617 20.5 3.94789 20.219 3.4478 19.719C2.9477 19.2189 2.66675 18.5406 2.66675 17.8333V6.16667C2.66675 5.45942 2.9477 4.78115 3.4478 4.28105Z"
+          fill="#999999"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const InfoIcon = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12 20C16.4183 20 20 16.4183 20 12C20 7.58172 16.4183 4 12 4C7.58172 4 4 7.58172 4 12C4 16.4183 7.58172 20 12 20ZM12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C6.47715 2 2 6.47715 2 12C2 17.5228 6.47715 22 12 22Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12.0083 16.6666C11.456 16.6666 11.0083 16.2189 11.0083 15.6666V11.3333C11.0083 10.781 11.456 10.3333 12.0083 10.3333C12.5606 10.3333 13.0083 10.781 13.0083 11.3333V15.6666C13.0083 16.2189 12.5606 16.6666 12.0083 16.6666Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M13.0083 8C13.0083 8.55228 12.5606 9 12.0083 9H12C11.4477 9 11 8.55228 11 8C11 7.44772 11.4477 7 12 7H12.0083C12.5606 7 13.0083 7.44772 13.0083 8Z"
+          fill="#1B1B1B"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const PaperClip = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M15.15 4.15625C14.5307 4.15625 13.9367 4.40228 13.4988 4.84021L5.84048 12.4985C5.08981 13.2492 4.66809 14.2673 4.66809 15.3289C4.66809 16.3905 5.08981 17.4087 5.84048 18.1593C6.59115 18.91 7.60927 19.3317 8.67087 19.3317C9.73247 19.3317 10.7506 18.91 11.5013 18.1593L19.1596 10.501C19.5501 10.1105 20.1833 10.1105 20.5738 10.501C20.9643 10.8915 20.9643 11.5247 20.5738 11.9152L12.9155 19.5735C11.7897 20.6993 10.2629 21.3317 8.67087 21.3317C7.07884 21.3317 5.552 20.6993 4.42626 19.5735C3.30052 18.4478 2.66809 16.921 2.66809 15.3289C2.66809 13.7369 3.30052 12.2101 4.42626 11.0843L12.0846 3.426C12.8976 2.61299 14.0003 2.15625 15.15 2.15625C16.2998 2.15625 17.4025 2.61299 18.2155 3.426C19.0285 4.239 19.4852 5.34167 19.4852 6.49144C19.4852 7.6412 19.0285 8.74387 18.2155 9.55688L10.5488 17.2152C10.0485 17.7155 9.37003 17.9965 8.66254 17.9965C7.95505 17.9965 7.27653 17.7155 6.77626 17.2152C6.27599 16.7149 5.99494 16.0364 5.99494 15.3289C5.99494 14.6214 6.27599 13.9429 6.77626 13.4427L13.8517 6.37558C14.2424 5.98529 14.8756 5.98566 15.2659 6.37641C15.6562 6.76717 15.6558 7.40033 15.2651 7.79063L8.19048 14.8569C8.19039 14.857 8.19057 14.8568 8.19048 14.8569C8.06545 14.982 7.99494 15.152 7.99494 15.3289C7.99494 15.506 8.06528 15.6758 8.19048 15.801C8.31568 15.9262 8.48548 15.9965 8.66254 15.9965C8.83959 15.9965 9.0094 15.9262 9.1346 15.801L16.8013 8.14266C17.239 7.70477 17.4852 7.11059 17.4852 6.49144C17.4852 5.87211 17.2392 5.27814 16.8013 4.84021C16.3633 4.40228 15.7694 4.15625 15.15 4.15625Z"
+          fill="#1B1B1B"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const People = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M9.49993 5.5C8.21126 5.5 7.1666 6.54467 7.1666 7.83333C7.1666 9.122 8.21126 10.1667 9.49993 10.1667C10.7886 10.1667 11.8333 9.122 11.8333 7.83333C11.8333 6.54467 10.7886 5.5 9.49993 5.5ZM5.1666 7.83333C5.1666 5.4401 7.10669 3.5 9.49993 3.5C11.8932 3.5 13.8333 5.4401 13.8333 7.83333C13.8333 10.2266 11.8932 12.1667 9.49993 12.1667C7.10669 12.1667 5.1666 10.2266 5.1666 7.83333Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M3.10246 14.7692C3.91511 13.9565 5.01731 13.5 6.16658 13.5H12.8333C13.9825 13.5 15.0847 13.9565 15.8974 14.7692C16.71 15.5819 17.1666 16.6841 17.1666 17.8333V19.5C17.1666 20.0523 16.7189 20.5 16.1666 20.5C15.6143 20.5 15.1666 20.0523 15.1666 19.5V17.8333C15.1666 17.2145 14.9208 16.621 14.4832 16.1834C14.0456 15.7458 13.4521 15.5 12.8333 15.5H6.16658C5.54775 15.5 4.95425 15.7458 4.51667 16.1834C4.07908 16.621 3.83325 17.2145 3.83325 17.8333V19.5C3.83325 20.0523 3.38554 20.5 2.83325 20.5C2.28097 20.5 1.83325 20.0523 1.83325 19.5V17.8333C1.83325 16.6841 2.2898 15.5819 3.10246 14.7692Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M14.3645 4.36029C14.5015 3.82527 15.0463 3.5026 15.5813 3.63958C16.5134 3.87824 17.3396 4.42034 17.9296 5.18042C18.5196 5.9405 18.8398 6.87532 18.8398 7.8375C18.8398 8.79969 18.5196 9.7345 17.9296 10.4946C17.3396 11.2547 16.5134 11.7968 15.5813 12.0354C15.0463 12.1724 14.5015 11.8497 14.3645 11.3147C14.2275 10.7797 14.5502 10.2349 15.0852 10.0979C15.5871 9.96941 16.032 9.67751 16.3497 9.26824C16.6674 8.85896 16.8398 8.3556 16.8398 7.8375C16.8398 7.3194 16.6674 6.81604 16.3497 6.40676C16.032 5.99749 15.5871 5.70559 15.0852 5.57708C14.5502 5.4401 14.2275 4.89532 14.3645 4.36029Z"
+          fill="#999999"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M17.6983 14.3583C17.8364 13.8236 18.3818 13.502 18.9166 13.6401C19.8462 13.8801 20.6698 14.4221 21.2581 15.1809C21.8463 15.9398 22.1659 16.8725 22.1666 17.8326V19.5C22.1666 20.0523 21.7189 20.5 21.1666 20.5C20.6143 20.5 20.1666 20.0523 20.1666 19.5V17.8338C20.1661 17.3169 19.9941 16.8148 19.6774 16.4063C19.3606 15.9977 18.9172 15.7058 18.4166 15.5766C17.8818 15.4385 17.5603 14.8931 17.6983 14.3583Z"
+          fill="#999999"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const PhoneIcon = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M19.92 20.5999H19.66C17.0438 20.313 14.5246 19.446 12.2861 18.062C10.0475 16.678 8.14591 14.812 6.71996 12.5999C5.22961 10.2865 4.30152 7.65636 4.00996 4.91994C3.97695 4.56982 4.01343 4.21661 4.11731 3.88063C4.22119 3.54464 4.39042 3.23248 4.61528 2.96209C4.84015 2.69169 5.11622 2.46838 5.42764 2.30498C5.73905 2.14158 6.07968 2.04132 6.42996 2.00994H9.15996C9.81089 2.00242 10.4418 2.23476 10.9324 2.66265C11.423 3.09054 11.7389 3.68403 11.82 4.32994C11.9197 5.05691 12.0973 5.77103 12.35 6.45994C12.5295 6.93661 12.5684 7.4548 12.462 7.95293C12.3557 8.45105 12.1085 8.90817 11.75 9.26994L11.24 9.77994C12.1834 11.1946 13.4048 12.4025 14.83 13.3299L15.33 12.8199C15.694 12.4616 16.1528 12.2147 16.6524 12.1084C17.152 12.0021 17.6716 12.0408 18.15 12.2199C18.8358 12.4714 19.5465 12.6491 20.27 12.7499C20.9175 12.836 21.5104 13.1582 21.9349 13.6547C22.3593 14.1512 22.5855 14.7869 22.57 15.4399V17.9299C22.5726 18.2796 22.506 18.6264 22.374 18.9502C22.242 19.274 22.0472 19.5685 21.8009 19.8167C21.5546 20.0649 21.2615 20.2619 20.9387 20.3963C20.6159 20.5308 20.2697 20.6 19.92 20.5999ZM9.17996 3.99994H6.65996C6.56889 4.00246 6.47925 4.02326 6.39637 4.06109C6.3135 4.09893 6.23906 4.15303 6.1775 4.22019C6.11594 4.28735 6.0685 4.36619 6.038 4.45204C6.0075 4.53789 5.99457 4.629 5.99996 4.71994C6.26463 7.13208 7.08624 9.4497 8.39996 11.4899C9.60251 13.3827 11.2072 14.9874 13.1 16.1899C15.1346 17.5094 17.4495 18.3347 19.86 18.5999C19.9513 18.6054 20.0427 18.5921 20.1287 18.5609C20.2146 18.5297 20.2933 18.4812 20.3598 18.4184C20.4263 18.3557 20.4793 18.2799 20.5154 18.1959C20.5516 18.1119 20.5701 18.0214 20.57 17.9299V15.3999C20.5739 15.2377 20.518 15.0797 20.4129 14.9561C20.3077 14.8325 20.1607 14.752 20 14.7299C19.1287 14.6126 18.2734 14.3979 17.45 14.0899C17.3294 14.0465 17.1991 14.0376 17.0736 14.0641C16.9482 14.0906 16.8327 14.1515 16.74 14.2399L15.69 15.3299C15.5349 15.4841 15.3337 15.5833 15.1169 15.6122C14.9002 15.6411 14.68 15.5981 14.49 15.4899C12.2467 14.2114 10.3885 12.3532 9.10996 10.1099C8.99053 9.92005 8.93841 9.69552 8.96199 9.47243C8.98557 9.24935 9.08347 9.04067 9.23996 8.87994L10.33 7.81994C10.4186 7.72939 10.4797 7.61556 10.5063 7.49166C10.5328 7.36775 10.5237 7.23886 10.48 7.11994C10.1732 6.29273 9.95854 5.43421 9.83996 4.55994C9.81588 4.40275 9.73584 4.25953 9.61458 4.15664C9.49332 4.05375 9.33898 3.99811 9.17996 3.99994Z"
+          fill="#242424"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const PlusBox = ({
+  setCreateType,
+  setIsCreating,
+  setIsEditing,
+  type,
+}) => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="32"
+        height="32"
+        viewBox="0 0 32 32"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+        onClick={() => {
+          setCreateType(type);
+          setIsCreating(prevState => !prevState);
+          setIsEditing(false);
+        }}
+      >
+        <path
+          d="M0 16C0 7.16344 7.16344 0 16 0V0C24.8366 0 32 7.16344 32 16V16C32 24.8366 24.8366 32 16 32V32C7.16344 32 0 24.8366 0 16V16Z"
+          fill="white"
+          fillOpacity="0.01"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M9.16667 8.5C8.98986 8.5 8.82029 8.57024 8.69526 8.69526C8.57024 8.82029 8.5 8.98986 8.5 9.16667V22.0858L10.1262 20.4596C10.3138 20.272 10.5681 20.1667 10.8333 20.1667H22.8333C23.0101 20.1667 23.1797 20.0964 23.3047 19.9714C23.4298 19.8464 23.5 19.6768 23.5 19.5V9.16667C23.5 8.98986 23.4298 8.82029 23.3047 8.69526C23.1797 8.57024 23.0101 8.5 22.8333 8.5H9.16667ZM7.28105 7.28105C7.78115 6.78095 8.45942 6.5 9.16667 6.5H22.8333C23.5406 6.5 24.2189 6.78095 24.719 7.28105C25.219 7.78115 25.5 8.45942 25.5 9.16667V19.5C25.5 20.2072 25.219 20.8855 24.719 21.3856C24.2189 21.8857 23.5406 22.1667 22.8333 22.1667H11.2475L8.20711 25.2071C7.92111 25.4931 7.49099 25.5787 7.11732 25.4239C6.74364 25.2691 6.5 24.9045 6.5 24.5V9.16667C6.5 8.45942 6.78095 7.78115 7.28105 7.28105Z"
+          fill="#242424"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M11.68 14.3333C11.68 13.781 12.1277 13.3333 12.68 13.3333H19.3467C19.899 13.3333 20.3467 13.781 20.3467 14.3333C20.3467 14.8855 19.899 15.3333 19.3467 15.3333H12.68C12.1277 15.3333 11.68 14.8855 11.68 14.3333Z"
+          fill="#242424"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M16.0133 10C16.5656 10 17.0133 10.4477 17.0133 11V17.6667C17.0133 18.219 16.5656 18.6667 16.0133 18.6667C15.4611 18.6667 15.0133 18.219 15.0133 17.6667V11C15.0133 10.4477 15.4611 10 16.0133 10Z"
+          fill="#242424"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const SearchIcon = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M11.3333 6.79922C8.82959 6.79922 6.79995 8.82886 6.79995 11.3326C6.79995 13.8362 8.82959 15.8659 11.3333 15.8659C13.837 15.8659 15.8666 13.8362 15.8666 11.3326C15.8666 8.82886 13.837 6.79922 11.3333 6.79922ZM5.19995 11.3326C5.19995 7.94521 7.94594 5.19922 11.3333 5.19922C14.7206 5.19922 17.4666 7.94521 17.4666 11.3326C17.4666 14.7199 14.7206 17.4659 11.3333 17.4659C7.94594 17.4659 5.19995 14.7199 5.19995 11.3326Z"
+          fill="black"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M14.5344 14.5351C14.8468 14.2227 15.3533 14.2227 15.6657 14.5351L18.5657 17.4351C18.8782 17.7475 18.8782 18.254 18.5657 18.5665C18.2533 18.8789 17.7468 18.8789 17.4344 18.5665L14.5344 15.6665C14.2219 15.354 14.2219 14.8475 14.5344 14.5351Z"
+          fill="black"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const SearchIconDash = () => {
+  return (
+    <div className="list__icon search">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M20.21 18.7892L17.21 15.7892C18.2438 14.4561 18.8033 12.8162 18.8 11.1292C18.8019 9.61138 18.3535 8.12709 17.5115 6.86423C16.6695 5.60138 15.4717 4.61674 14.0698 4.03498C12.6678 3.45322 11.1248 3.30048 9.63607 3.59611C8.14731 3.89174 6.77973 4.62244 5.70646 5.69571C4.63318 6.76898 3.90248 8.13656 3.60685 9.62533C3.31122 11.1141 3.46396 12.6571 4.04572 14.059C4.62749 15.4609 5.61212 16.6587 6.87497 17.5008C8.13783 18.3428 9.62213 18.7912 11.14 18.7892C12.8269 18.7925 14.4668 18.233 15.8 17.1992L18.8 20.1992C18.9883 20.3875 19.2437 20.4933 19.51 20.4933C19.7763 20.4933 20.0317 20.3875 20.22 20.1992C20.4083 20.0109 20.514 19.7555 20.514 19.4892C20.514 19.2229 20.4083 18.9675 20.22 18.7792L20.21 18.7892ZM5.49996 11.1692C5.50185 10.2385 5.73282 9.32258 6.17248 8.50227C6.61214 7.68196 7.24697 6.98248 8.02095 6.46559C8.79492 5.94869 9.68423 5.63027 10.6104 5.53842C11.5366 5.44657 12.4711 5.58413 13.3315 5.93894C14.1919 6.29375 14.9518 6.85491 15.544 7.5729C16.1362 8.29088 16.5426 9.1436 16.7272 10.0558C16.9119 10.968 16.8691 11.9116 16.6028 12.8034C16.3364 13.6952 15.8546 14.5077 15.2 15.1692C14.4061 15.9517 13.3992 16.4828 12.3051 16.6961C11.2111 16.9095 10.0784 16.7956 9.0487 16.3688C8.01899 15.942 7.13793 15.2211 6.51565 14.2963C5.89337 13.3715 5.50748 12.2839 5.49996 11.1692Z"
+          fill="#1B1B1B"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const Settings = () => {
+  return (
+    <div className="nav__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
+        <path
+          d="M18.9799 14.8482C19.0161 14.7746 19.0745 14.7141 19.1468 14.6753C19.2192 14.6365 19.3018 14.6213 19.3832 14.6318C20.0798 14.6318 20.7479 14.3551 21.2405 13.8625C21.7331 13.3699 22.0099 12.7018 22.0099 12.0051C22.0099 11.3085 21.7331 10.6404 21.2405 10.1478C20.7479 9.65521 20.0798 9.37847 19.3832 9.37847H19.2455C19.1906 9.3785 19.1365 9.36612 19.0871 9.34228C19.0377 9.31843 18.9943 9.28372 18.9602 9.24075C18.9497 9.17453 18.9263 9.11102 18.8913 9.05383C18.8615 8.97548 18.8569 8.88981 18.8781 8.80871C18.8992 8.72761 18.9451 8.65511 19.0094 8.6013C19.2729 8.36026 19.4845 8.06802 19.6312 7.74243C19.778 7.41685 19.8568 7.06475 19.8628 6.70768C19.8689 6.3506 19.8021 5.99604 19.6665 5.66567C19.5308 5.33529 19.3293 5.03604 19.0741 4.78621C18.8189 4.53637 18.5154 4.3412 18.1822 4.21262C17.8491 4.08404 17.4932 4.02476 17.1363 4.03839C16.7794 4.05202 16.4291 4.13828 16.1067 4.2919C15.7843 4.44552 15.4966 4.66327 15.2612 4.93184C15.2063 4.98531 15.1362 5.0206 15.0606 5.03292C14.9849 5.04523 14.9073 5.03399 14.8382 5.00071C14.7646 4.96445 14.7041 4.90607 14.6653 4.83374C14.6265 4.76142 14.6113 4.67877 14.6218 4.59736C14.6218 3.90073 14.345 3.23263 13.8524 2.74003C13.3598 2.24744 12.6917 1.9707 11.9951 1.9707C11.2985 1.9707 10.6304 2.24744 10.1378 2.74003C9.64519 3.23263 9.36845 3.90073 9.36845 4.59736V4.73509C9.36848 4.78995 9.3561 4.8441 9.33225 4.89351C9.3084 4.94291 9.27369 4.98628 9.23072 5.02038C9.16573 5.04104 9.10552 5.07449 9.05364 5.11876C8.9753 5.14854 8.88962 5.15317 8.80852 5.13202C8.72742 5.11086 8.65493 5.06496 8.60111 5.00071C8.36738 4.71296 8.07539 4.47794 7.74433 4.31112C7.41327 4.14429 7.05065 4.04943 6.68031 4.03278C6.30997 4.01612 5.94029 4.07805 5.5956 4.21448C5.2509 4.35092 4.93899 4.55877 4.68037 4.82438C4.42176 5.08999 4.22229 5.40733 4.09509 5.75554C3.9679 6.10375 3.91585 6.47495 3.94237 6.84471C3.96889 7.21448 4.07339 7.57444 4.24898 7.90093C4.42458 8.22742 4.66729 8.51305 4.96117 8.73903C5.01765 8.80051 5.05366 8.878 5.06424 8.96082C5.07481 9.04364 5.05942 9.12769 5.0202 9.2014C4.98914 9.2788 4.93428 9.34434 4.86355 9.38854C4.79283 9.43275 4.70988 9.45334 4.62669 9.44734C4.26889 9.42541 3.91039 9.47695 3.57326 9.59879C3.23613 9.72063 2.9275 9.91018 2.66639 10.1558C2.40527 10.4014 2.1972 10.6978 2.05496 11.0269C1.91273 11.3559 1.83936 11.7106 1.83936 12.0691C1.83936 12.4275 1.91273 12.7822 2.05496 13.1113C2.1972 13.4403 2.40527 13.7368 2.66639 13.9824C2.9275 14.228 3.23613 14.4175 3.57326 14.5394C3.91039 14.6612 4.26889 14.7127 4.62669 14.6908H4.76442C4.84484 14.6883 4.92374 14.7132 4.98811 14.7615C5.05248 14.8098 5.0985 14.8785 5.11857 14.9564C5.1526 15.0333 5.16044 15.1192 5.14089 15.201C5.12134 15.2827 5.07548 15.3558 5.01036 15.409C4.74684 15.65 4.53524 15.9422 4.3885 16.2678C4.24176 16.5934 4.16295 16.9455 4.15689 17.3026C4.15084 17.6597 4.21766 18.0142 4.35328 18.3446C4.4889 18.675 4.69046 18.9742 4.94565 19.2241C5.20085 19.4739 5.50431 19.6691 5.83748 19.7976C6.17066 19.9262 6.52656 19.9855 6.88343 19.9719C7.2403 19.9582 7.59064 19.872 7.91304 19.7184C8.23544 19.5647 8.52312 19.347 8.75851 19.0784C8.82191 19.0213 8.90177 18.9858 8.98661 18.977C9.07146 18.9682 9.15693 18.9864 9.23072 19.0292C9.30644 19.0588 9.37045 19.1123 9.41306 19.1815C9.45566 19.2507 9.47454 19.332 9.46683 19.4129C9.4449 19.7707 9.49644 20.1292 9.61828 20.4663C9.74012 20.8035 9.92967 21.1121 10.1753 21.3732C10.4209 21.6343 10.7173 21.8424 11.0464 21.9846C11.3754 22.1269 11.7301 22.2002 12.0886 22.2002C12.447 22.2002 12.8017 22.1269 13.1308 21.9846C13.4598 21.8424 13.7563 21.6343 14.0019 21.3732C14.2475 21.1121 14.437 20.8035 14.5589 20.4663C14.6807 20.1292 14.7322 19.7707 14.7103 19.4129V19.2752C14.7049 19.1933 14.7259 19.1119 14.7701 19.0429C14.8144 18.9738 14.8796 18.9208 14.9563 18.8915C15.0327 18.8654 15.115 18.8617 15.1934 18.8809C15.2719 18.9001 15.3432 18.9413 15.3989 18.9997C15.6401 19.245 15.9281 19.4393 16.2458 19.5711C16.5635 19.7029 16.9045 19.7696 17.2484 19.7671C17.594 19.7709 17.9369 19.7049 18.2564 19.573C18.5759 19.4412 18.8654 19.2462 19.1078 18.9997C19.3555 18.7546 19.5522 18.4627 19.6865 18.141C19.8207 17.8193 19.8898 17.4742 19.8898 17.1256C19.8898 16.7771 19.8207 16.432 19.6865 16.1103C19.5522 15.7886 19.3555 15.4967 19.1078 15.2516C19.0558 15.1978 19.0207 15.1301 19.0067 15.0567C18.9928 14.9833 19.0005 14.9074 19.0291 14.8384L18.9799 14.8482ZM17.1697 14.071C16.979 14.5099 16.9268 14.9965 17.0199 15.4659C17.1131 15.9352 17.3473 16.365 17.6911 16.6977C17.7539 16.7583 17.8038 16.8309 17.8379 16.9111C17.872 16.9914 17.8896 17.0778 17.8896 17.165C17.8896 17.2522 17.872 17.3386 17.8379 17.4188C17.8038 17.4991 17.7539 17.5717 17.6911 17.6323C17.5674 17.7527 17.4014 17.8199 17.2288 17.8192C17.0563 17.8188 16.8907 17.7519 16.7664 17.6323L16.7074 17.5733C16.3761 17.2462 15.9541 17.0263 15.4962 16.9421C15.0383 16.8579 14.5657 16.9133 14.1397 17.1011C13.7209 17.2825 13.364 17.5818 13.1124 17.9627C12.8609 18.3435 12.7256 18.7893 12.7231 19.2457V19.3834C12.7231 19.5565 12.655 19.7227 12.5335 19.846C12.412 19.9694 12.2469 20.0399 12.0738 20.0425C11.9798 20.039 11.8875 20.0162 11.8026 19.9756C11.7177 19.935 11.6421 19.8774 11.5803 19.8064C11.5186 19.7354 11.472 19.6525 11.4435 19.5629C11.415 19.4732 11.4052 19.3786 11.4147 19.285C11.4093 18.8251 11.2657 18.3775 11.0025 18.0002C10.7394 17.623 10.3688 17.3337 9.93904 17.1699C9.64746 17.0419 9.33274 16.9749 9.01429 16.9732C8.37479 16.9802 7.76363 17.2381 7.31238 17.6913C7.25075 17.7539 7.17723 17.8036 7.09614 17.8374C7.01505 17.8712 6.92802 17.8884 6.84017 17.8881C6.66707 17.8855 6.50193 17.8149 6.38044 17.6916C6.25895 17.5682 6.19086 17.4021 6.19088 17.2289C6.19422 17.1299 6.21775 17.0325 6.26004 16.9429C6.30233 16.8532 6.36249 16.7731 6.43682 16.7075C6.76389 16.3762 6.98381 15.9543 7.06801 15.4964C7.15222 15.0385 7.09683 14.5659 6.90903 14.1399C6.72762 13.7211 6.42825 13.3642 6.04743 13.1126C5.66661 12.8611 5.22081 12.7258 4.76442 12.7233H4.62669C4.45357 12.7233 4.28739 12.6552 4.16406 12.5337C4.04073 12.4122 3.97015 12.2471 3.96756 12.074C3.97112 11.98 3.9939 11.8877 4.03451 11.8028C4.07512 11.7179 4.13269 11.6423 4.20367 11.5805C4.27466 11.5187 4.35754 11.4722 4.44722 11.4437C4.53689 11.4152 4.63145 11.4054 4.72507 11.4149C5.18497 11.4095 5.63263 11.2659 6.00984 11.0027C6.38704 10.7396 6.67636 10.369 6.84017 9.93922C7.0325 9.49992 7.08469 9.01199 6.98963 8.54194C6.89456 8.0719 6.65687 7.6426 6.30893 7.31256C6.18239 7.18725 6.10849 7.01834 6.10234 6.84035C6.10493 6.66725 6.17551 6.50212 6.29884 6.38063C6.42217 6.25914 6.58835 6.19105 6.76147 6.19107C6.86054 6.1944 6.95789 6.21794 7.04755 6.26023C7.1372 6.30252 7.21728 6.36267 7.28286 6.43701C7.59056 6.7422 7.97785 6.95475 8.40054 7.05042C8.82322 7.14609 9.26429 7.12103 9.67342 6.97808C9.7545 6.96426 9.83372 6.94115 9.90952 6.90922C10.3298 6.7299 10.6882 6.43103 10.94 6.04974C11.1919 5.66846 11.3262 5.22156 11.3261 4.7646V4.62688C11.3261 4.53864 11.3438 4.45129 11.3781 4.37001C11.4125 4.28873 11.4628 4.21518 11.5261 4.15371C11.5894 4.09224 11.6644 4.04411 11.7467 4.01216C11.8289 3.98022 11.9167 3.96512 12.0049 3.96775C12.0976 3.97008 12.1888 3.99137 12.2729 4.0303C12.3571 4.06923 12.4323 4.12498 12.4941 4.19411C12.5558 4.26324 12.6028 4.34428 12.632 4.43225C12.6613 4.52021 12.6722 4.61324 12.6641 4.70558C12.6639 5.16004 12.7971 5.60457 13.0472 5.98405C13.2972 6.36352 13.6532 6.66126 14.0709 6.84035C14.5107 7.02985 14.9979 7.08052 15.4673 6.98558C15.9367 6.89065 16.3659 6.65463 16.6975 6.30912C16.8237 6.18941 16.9909 6.12267 17.1648 6.12267C17.3387 6.12267 17.506 6.18941 17.6321 6.30912C17.7526 6.43281 17.8197 6.59883 17.819 6.77149C17.8157 6.87056 17.7921 6.96791 17.7499 7.05757C17.7076 7.14723 17.6474 7.2273 17.5731 7.29289C17.2699 7.60184 17.0586 7.98911 16.9631 8.41133C16.8675 8.83354 16.8914 9.27403 17.032 9.68344C17.0458 9.76452 17.0689 9.84375 17.1009 9.91955C17.2802 10.3399 17.5791 10.6982 17.9603 10.9501C18.3416 11.2019 18.7885 11.3362 19.2455 11.3362H19.3832C19.558 11.3362 19.7257 11.4056 19.8493 11.5292C19.9729 11.6528 20.0423 11.8205 20.0423 11.9953C20.04 12.088 20.0187 12.1792 19.9798 12.2633C19.9408 12.3474 19.8851 12.4227 19.816 12.4844C19.7468 12.5462 19.6658 12.5931 19.5778 12.6224C19.4899 12.6516 19.3968 12.6625 19.3045 12.6544C18.8489 12.6551 18.4036 12.7897 18.024 13.0416C17.6444 13.2935 17.3473 13.6515 17.1697 14.071Z"
+          fill="#999999"
+        />
+        <path
+          d="M12.0049 8.56195C11.3239 8.56195 10.6582 8.76389 10.092 9.14223C9.52579 9.52057 9.08446 10.0583 8.82386 10.6875C8.56325 11.3166 8.49506 12.009 8.62792 12.6769C8.76078 13.3448 9.08871 13.9583 9.57025 14.4398C10.0518 14.9214 10.6653 15.2493 11.3332 15.3822C12.0011 15.515 12.6934 15.4468 13.3226 15.1862C13.9518 14.9256 14.4895 14.4843 14.8679 13.9181C15.2462 13.3518 15.4481 12.6861 15.4481 12.0051C15.4481 11.0919 15.0854 10.2162 14.4396 9.57043C13.7939 8.92471 12.9181 8.56195 12.0049 8.56195ZM12.0049 13.4808C11.7131 13.4808 11.4278 13.3942 11.1851 13.2321C10.9425 13.0699 10.7533 12.8395 10.6416 12.5698C10.5299 12.3002 10.5007 12.0035 10.5577 11.7172C10.6146 11.431 10.7551 11.1681 10.9615 10.9617C11.1679 10.7553 11.4308 10.6148 11.7171 10.5578C12.0033 10.5009 12.3 10.5301 12.5697 10.6418C12.8393 10.7535 13.0698 10.9426 13.2319 11.1853C13.3941 11.428 13.4806 11.7133 13.4806 12.0051C13.4806 12.3965 13.3251 12.7718 13.0484 13.0486C12.7717 13.3253 12.3963 13.4808 12.0049 13.4808Z"
+          fill="#999999"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const UserPlus = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M18.6667 14.334C19.219 14.334 19.6667 14.7817 19.6667 15.334V19.5007C19.6667 20.0529 19.219 20.5007 18.6667 20.5007C18.1145 20.5007 17.6667 20.0529 17.6667 19.5007V15.334C17.6667 14.7817 18.1145 14.334 18.6667 14.334Z"
+          fill="#242424"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M15.1667 15.334C15.1667 14.7817 15.6145 14.334 16.1667 14.334H21.1667C21.719 14.334 22.1667 14.7817 22.1667 15.334C22.1667 15.8863 21.719 16.334 21.1667 16.334H16.1667C15.6145 16.334 15.1667 15.8863 15.1667 15.334Z"
+          fill="#242424"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M5.33325 12.666C5.88554 12.666 6.33325 13.1137 6.33325 13.666V19.4993C6.33325 20.0516 5.88554 20.4993 5.33325 20.4993C4.78097 20.4993 4.33325 20.0516 4.33325 19.4993V13.666C4.33325 13.1137 4.78097 12.666 5.33325 12.666Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M1.83325 13.666C1.83325 13.1137 2.28097 12.666 2.83325 12.666H7.83325C8.38554 12.666 8.83325 13.1137 8.83325 13.666C8.83325 14.2183 8.38554 14.666 7.83325 14.666H2.83325C2.28097 14.666 1.83325 14.2183 1.83325 13.666Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12 11C12.5523 11 13 11.4477 13 12V19.5C13 20.0523 12.5523 20.5 12 20.5C11.4477 20.5 11 20.0523 11 19.5V12C11 11.4477 11.4477 11 12 11Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M8.5 8.66602C8.5 8.11373 8.94772 7.66602 9.5 7.66602H14.5C15.0523 7.66602 15.5 8.11373 15.5 8.66602C15.5 9.2183 15.0523 9.66602 14.5 9.66602H9.5C8.94772 9.66602 8.5 9.2183 8.5 8.66602Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M18.6667 3.5C19.219 3.5 19.6667 3.94772 19.6667 4.5V12C19.6667 12.5523 19.219 13 18.6667 13C18.1145 13 17.6667 12.5523 17.6667 12V4.5C17.6667 3.94772 18.1145 3.5 18.6667 3.5Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M12 3.5C12.5523 3.5 13 3.94772 13 4.5V8.66667C13 9.21895 12.5523 9.66667 12 9.66667C11.4477 9.66667 11 9.21895 11 8.66667V4.5C11 3.94772 11.4477 3.5 12 3.5Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M5.33325 3.5C5.88554 3.5 6.33325 3.94772 6.33325 4.5V10.3333C6.33325 10.8856 5.88554 11.3333 5.33325 11.3333C4.78097 11.3333 4.33325 10.8856 4.33325 10.3333V4.5C4.33325 3.94772 4.78097 3.5 5.33325 3.5Z"
+          fill="#1B1B1B"
+        />
+      </svg>
+    </div>
+  );
+};
+
+export const Silders = () => {
+  return (
+    <div className="list__icon">
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M9.08325 5.5C7.79459 5.5 6.74992 6.54467 6.74992 7.83333C6.74992 9.122 7.79459 10.1667 9.08325 10.1667C10.3719 10.1667 11.4166 9.122 11.4166 7.83333C11.4166 6.54467 10.3719 5.5 9.08325 5.5ZM4.74992 7.83333C4.74992 5.4401 6.69002 3.5 9.08325 3.5C11.4765 3.5 13.4166 5.4401 13.4166 7.83333C13.4166 10.2266 11.4765 12.1667 9.08325 12.1667C6.69002 12.1667 4.74992 10.2266 4.74992 7.83333Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M15.1666 11.1667C15.1666 10.6144 15.6143 10.1667 16.1666 10.1667H21.1666C21.7189 10.1667 22.1666 10.6144 22.1666 11.1667C22.1666 11.7189 21.7189 12.1667 21.1666 12.1667H16.1666C15.6143 12.1667 15.1666 11.7189 15.1666 11.1667Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M18.6666 7.66666C19.2189 7.66666 19.6666 8.11437 19.6666 8.66666V13.6667C19.6666 14.2189 19.2189 14.6667 18.6666 14.6667C18.1143 14.6667 17.6666 14.2189 17.6666 13.6667V8.66666C17.6666 8.11437 18.1143 7.66666 18.6666 7.66666Z"
+          fill="#1B1B1B"
+        />
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M3.10246 14.7692C3.91511 13.9565 5.01731 13.5 6.16658 13.5H11.9999C13.1492 13.5 14.2514 13.9565 15.064 14.7692C15.8767 15.5819 16.3333 16.6841 16.3333 17.8333V19.5C16.3333 20.0523 15.8855 20.5 15.3333 20.5C14.781 20.5 14.3333 20.0523 14.3333 19.5V17.8333C14.3333 17.2145 14.0874 16.621 13.6498 16.1834C13.2122 15.7458 12.6188 15.5 11.9999 15.5H6.16658C5.54775 15.5 4.95425 15.7458 4.51667 16.1834C4.07908 16.621 3.83325 17.2145 3.83325 17.8333V19.5C3.83325 20.0523 3.38554 20.5 2.83325 20.5C2.28097 20.5 1.83325 20.0523 1.83325 19.5V17.8333C1.83325 16.6841 2.2898 15.5819 3.10246 14.7692Z"
+          fill="#1B1B1B"
+        />
+      </svg>
+    </div>
+  );
+};
+export const EmojiIcon = ({ openEmojiPicker }) => {
+  return (
+    <div className="list__icon" onClick={openEmojiPicker}>
+      <svg
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M11.9999 2.66992C10.1546 2.66992 8.35077 3.21712 6.81646 4.24231C5.28215 5.2675 4.0863 6.72465 3.38013 8.42949C2.67396 10.1343 2.4892 12.0103 2.8492 13.8201C3.2092 15.63 4.0978 17.2924 5.40262 18.5972C6.70744 19.9021 8.36989 20.7906 10.1797 21.1506C11.9896 21.5106 13.8655 21.3259 15.5704 20.6197C17.2752 19.9136 18.7323 18.7177 19.7575 17.1834C20.7827 15.6491 21.3299 13.8452 21.3299 11.9999C21.3299 9.52545 20.3469 7.15233 18.5972 5.40262C16.8475 3.6529 14.4744 2.66992 11.9999 2.66992ZM11.9999 19.3299C10.5502 19.3299 9.13301 18.9 7.9276 18.0946C6.72218 17.2892 5.78268 16.1444 5.22789 14.805C4.6731 13.4656 4.52794 11.9918 4.81077 10.5699C5.0936 9.14803 5.79171 7.84195 6.81683 6.81683C7.84195 5.79171 9.14803 5.0936 10.5699 4.81077C11.9918 4.52794 13.4656 4.67309 14.805 5.22788C16.1444 5.78268 17.2892 6.72218 18.0946 7.92759C18.9 9.133 19.3299 10.5502 19.3299 11.9999C19.3273 13.9431 18.5542 15.806 17.1801 17.1801C15.806 18.5542 13.9432 19.3273 11.9999 19.3299Z"
+          fill="#1B1B1B"
+        />
+        <path
+          d="M14.5399 13.0599C14.2272 13.4355 13.84 13.7422 13.4029 13.9608C12.9657 14.1794 12.4881 14.305 11.9999 14.3299C11.5118 14.305 11.0342 14.1794 10.597 13.9608C10.1599 13.7422 9.77271 13.4355 9.45994 13.0599C9.29506 12.862 9.06018 12.7354 8.80425 12.7064C8.54831 12.6774 8.29106 12.7483 8.08609 12.9043C7.88112 13.0603 7.74425 13.2893 7.70398 13.5437C7.66371 13.7981 7.72316 14.0583 7.86994 14.2699C8.36891 14.891 8.99641 15.3967 9.70932 15.7523C10.4222 16.1079 11.2037 16.3049 11.9999 16.3299C12.7962 16.3049 13.5777 16.1079 14.2906 15.7523C15.0035 15.3967 15.631 14.891 16.1299 14.2699C16.2164 14.1662 16.2809 14.046 16.3197 13.9166C16.3584 13.7873 16.3705 13.6514 16.3553 13.5172C16.3401 13.383 16.2979 13.2533 16.2312 13.1359C16.1645 13.0185 16.0747 12.9158 15.9673 12.834C15.8598 12.7522 15.7369 12.6931 15.606 12.6601C15.475 12.6271 15.3387 12.621 15.2054 12.6421C15.072 12.6632 14.9443 12.7111 14.8299 12.7829C14.7155 12.8547 14.6169 12.949 14.5399 13.0599Z"
+          fill="#1B1B1B"
+        />
+        <path
+          d="M14.5099 10.4999C14.7752 10.4999 15.0295 10.3946 15.2171 10.207C15.4046 10.0195 15.5099 9.76514 15.5099 9.49992C15.5099 9.23471 15.4046 8.98035 15.2171 8.79282C15.0295 8.60528 14.7752 8.49992 14.5099 8.49992C14.2447 8.49992 13.9904 8.60528 13.8028 8.79282C13.6153 8.98035 13.5099 9.23471 13.5099 9.49992C13.5099 9.76514 13.6153 10.0195 13.8028 10.207C13.9904 10.3946 14.2447 10.4999 14.5099 10.4999Z"
+          fill="#1B1B1B"
+        />
+        <path
+          d="M9.50995 10.4999C9.77517 10.4999 10.0295 10.3946 10.2171 10.207C10.4046 10.0195 10.5099 9.76514 10.5099 9.49992C10.5099 9.23471 10.4046 8.98035 10.2171 8.79282C10.0295 8.60528 9.77517 8.49992 9.50995 8.49992C9.24473 8.49992 8.99038 8.60528 8.80284 8.79282C8.61531 8.98035 8.50995 9.23471 8.50995 9.49992C8.50995 9.76514 8.61531 10.0195 8.80284 10.207C8.99038 10.3946 9.24473 10.4999 9.50995 10.4999Z"
+          fill="#1B1B1B"
         />
       </svg>
     </div>
