--- src/assets/InfoIcon.js
@@ -1,34 +1,0 @@
-import React from 'react';
-
-export const InfoIcon = () => {
-  return (
-    <div className="list__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12 20C16.4183 20 20 16.4183 20 12C20 7.58172 16.4183 4 12 4C7.58172 4 4 7.58172 4 12C4 16.4183 7.58172 20 12 20ZM12 22C17.5228 22 22 17.5228 22 12C22 6.47715 17.5228 2 12 2C6.47715 2 2 6.47715 2 12C2 17.5228 6.47715 22 12 22Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12.0083 16.6666C11.456 16.6666 11.0083 16.2189 11.0083 15.6666V11.3333C11.0083 10.781 11.456 10.3333 12.0083 10.3333C12.5606 10.3333 13.0083 10.781 13.0083 11.3333V15.6666C13.0083 16.2189 12.5606 16.6666 12.0083 16.6666Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M13.0083 8C13.0083 8.55228 12.5606 9 12.0083 9H12C11.4477 9 11 8.55228 11 8C11 7.44772 11.4477 7 12 7H12.0083C12.5606 7 13.0083 7.44772 13.0083 8Z"
-          fill="#1B1B1B"
-        />
-      </svg>
-    </div>
-  );
-};
