--- src/assets/People.js
@@ -1,42 +1,0 @@
-import React from 'react';
-
-const People = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M9.49993 5.5C8.21126 5.5 7.1666 6.54467 7.1666 7.83333C7.1666 9.122 8.21126 10.1667 9.49993 10.1667C10.7886 10.1667 11.8333 9.122 11.8333 7.83333C11.8333 6.54467 10.7886 5.5 9.49993 5.5ZM5.1666 7.83333C5.1666 5.4401 7.10669 3.5 9.49993 3.5C11.8932 3.5 13.8333 5.4401 13.8333 7.83333C13.8333 10.2266 11.8932 12.1667 9.49993 12.1667C7.10669 12.1667 5.1666 10.2266 5.1666 7.83333Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M3.10246 14.7692C3.91511 13.9565 5.01731 13.5 6.16658 13.5H12.8333C13.9825 13.5 15.0847 13.9565 15.8974 14.7692C16.71 15.5819 17.1666 16.6841 17.1666 17.8333V19.5C17.1666 20.0523 16.7189 20.5 16.1666 20.5C15.6143 20.5 15.1666 20.0523 15.1666 19.5V17.8333C15.1666 17.2145 14.9208 16.621 14.4832 16.1834C14.0456 15.7458 13.4521 15.5 12.8333 15.5H6.16658C5.54775 15.5 4.95425 15.7458 4.51667 16.1834C4.07908 16.621 3.83325 17.2145 3.83325 17.8333V19.5C3.83325 20.0523 3.38554 20.5 2.83325 20.5C2.28097 20.5 1.83325 20.0523 1.83325 19.5V17.8333C1.83325 16.6841 2.2898 15.5819 3.10246 14.7692Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M14.3645 4.36029C14.5015 3.82527 15.0463 3.5026 15.5813 3.63958C16.5134 3.87824 17.3396 4.42034 17.9296 5.18042C18.5196 5.9405 18.8398 6.87532 18.8398 7.8375C18.8398 8.79969 18.5196 9.7345 17.9296 10.4946C17.3396 11.2547 16.5134 11.7968 15.5813 12.0354C15.0463 12.1724 14.5015 11.8497 14.3645 11.3147C14.2275 10.7797 14.5502 10.2349 15.0852 10.0979C15.5871 9.96941 16.032 9.67751 16.3497 9.26824C16.6674 8.85896 16.8398 8.3556 16.8398 7.8375C16.8398 7.3194 16.6674 6.81604 16.3497 6.40676C16.032 5.99749 15.5871 5.70559 15.0852 5.57708C14.5502 5.4401 14.2275 4.89532 14.3645 4.36029Z"
-          fill="#999999"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M17.6983 14.3583C17.8364 13.8236 18.3818 13.502 18.9166 13.6401C19.8462 13.8801 20.6698 14.4221 21.2581 15.1809C21.8463 15.9398 22.1659 16.8725 22.1666 17.8326V19.5C22.1666 20.0523 21.7189 20.5 21.1666 20.5C20.6143 20.5 20.1666 20.0523 20.1666 19.5V17.8338C20.1661 17.3169 19.9941 16.8148 19.6774 16.4063C19.3606 15.9977 18.9172 15.7058 18.4166 15.5766C17.8818 15.4385 17.5603 14.8931 17.6983 14.3583Z"
-          fill="#999999"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default People;
