--- src/assets/PhoneIcon.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-export const PhoneIcon = () => {
-  return (
-    <div className="list__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          d="M19.92 20.5999H19.66C17.0438 20.313 14.5246 19.446 12.2861 18.062C10.0475 16.678 8.14591 14.812 6.71996 12.5999C5.22961 10.2865 4.30152 7.65636 4.00996 4.91994C3.97695 4.56982 4.01343 4.21661 4.11731 3.88063C4.22119 3.54464 4.39042 3.23248 4.61528 2.96209C4.84015 2.69169 5.11622 2.46838 5.42764 2.30498C5.73905 2.14158 6.07968 2.04132 6.42996 2.00994H9.15996C9.81089 2.00242 10.4418 2.23476 10.9324 2.66265C11.423 3.09054 11.7389 3.68403 11.82 4.32994C11.9197 5.05691 12.0973 5.77103 12.35 6.45994C12.5295 6.93661 12.5684 7.4548 12.462 7.95293C12.3557 8.45105 12.1085 8.90817 11.75 9.26994L11.24 9.77994C12.1834 11.1946 13.4048 12.4025 14.83 13.3299L15.33 12.8199C15.694 12.4616 16.1528 12.2147 16.6524 12.1084C17.152 12.0021 17.6716 12.0408 18.15 12.2199C18.8358 12.4714 19.5465 12.6491 20.27 12.7499C20.9175 12.836 21.5104 13.1582 21.9349 13.6547C22.3593 14.1512 22.5855 14.7869 22.57 15.4399V17.9299C22.5726 18.2796 22.506 18.6264 22.374 18.9502C22.242 19.274 22.0472 19.5685 21.8009 19.8167C21.5546 20.0649 21.2615 20.2619 20.9387 20.3963C20.6159 20.5308 20.2697 20.6 19.92 20.5999ZM9.17996 3.99994H6.65996C6.56889 4.00246 6.47925 4.02326 6.39637 4.06109C6.3135 4.09893 6.23906 4.15303 6.1775 4.22019C6.11594 4.28735 6.0685 4.36619 6.038 4.45204C6.0075 4.53789 5.99457 4.629 5.99996 4.71994C6.26463 7.13208 7.08624 9.4497 8.39996 11.4899C9.60251 13.3827 11.2072 14.9874 13.1 16.1899C15.1346 17.5094 17.4495 18.3347 19.86 18.5999C19.9513 18.6054 20.0427 18.5921 20.1287 18.5609C20.2146 18.5297 20.2933 18.4812 20.3598 18.4184C20.4263 18.3557 20.4793 18.2799 20.5154 18.1959C20.5516 18.1119 20.5701 18.0214 20.57 17.9299V15.3999C20.5739 15.2377 20.518 15.0797 20.4129 14.9561C20.3077 14.8325 20.1607 14.752 20 14.7299C19.1287 14.6126 18.2734 14.3979 17.45 14.0899C17.3294 14.0465 17.1991 14.0376 17.0736 14.0641C16.9482 14.0906 16.8327 14.1515 16.74 14.2399L15.69 15.3299C15.5349 15.4841 15.3337 15.5833 15.1169 15.6122C14.9002 15.6411 14.68 15.5981 14.49 15.4899C12.2467 14.2114 10.3885 12.3532 9.10996 10.1099C8.99053 9.92005 8.93841 9.69552 8.96199 9.47243C8.98557 9.24935 9.08347 9.04067 9.23996 8.87994L10.33 7.81994C10.4186 7.72939 10.4797 7.61556 10.5063 7.49166C10.5328 7.36775 10.5237 7.23886 10.48 7.11994C10.1732 6.29273 9.95854 5.43421 9.83996 4.55994C9.81588 4.40275 9.73584 4.25953 9.61458 4.15664C9.49332 4.05375 9.33898 3.99811 9.17996 3.99994Z"
-          fill="#242424"
-        />
-      </svg>
-    </div>
-  );
-};
