--- src/assets/PlusBox.js
@@ -1,51 +1,0 @@
-import React from 'react';
-
-export const PlusBox = ({
-  setCreateType,
-  setIsCreating,
-  setIsEditing,
-  type,
-}) => {
-  return (
-    <div className="list__icon">
-      <svg
-        width="32"
-        height="32"
-        viewBox="0 0 32 32"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-        onClick={() => {
-          setCreateType(type);
-          setIsCreating(prevState => !prevState);
-          setIsEditing(false);
-        }}
-      >
-        <path
-          d="M0 16C0 7.16344 7.16344 0 16 0V0C24.8366 0 32 7.16344 32 16V16C32 24.8366 24.8366 32 16 32V32C7.16344 32 0 24.8366 0 16V16Z"
-          fill="white"
-          fillOpacity="0.01"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M9.16667 8.5C8.98986 8.5 8.82029 8.57024 8.69526 8.69526C8.57024 8.82029 8.5 8.98986 8.5 9.16667V22.0858L10.1262 20.4596C10.3138 20.272 10.5681 20.1667 10.8333 20.1667H22.8333C23.0101 20.1667 23.1797 20.0964 23.3047 19.9714C23.4298 19.8464 23.5 19.6768 23.5 19.5V9.16667C23.5 8.98986 23.4298 8.82029 23.3047 8.69526C23.1797 8.57024 23.0101 8.5 22.8333 8.5H9.16667ZM7.28105 7.28105C7.78115 6.78095 8.45942 6.5 9.16667 6.5H22.8333C23.5406 6.5 24.2189 6.78095 24.719 7.28105C25.219 7.78115 25.5 8.45942 25.5 9.16667V19.5C25.5 20.2072 25.219 20.8855 24.719 21.3856C24.2189 21.8857 23.5406 22.1667 22.8333 22.1667H11.2475L8.20711 25.2071C7.92111 25.4931 7.49099 25.5787 7.11732 25.4239C6.74364 25.2691 6.5 24.9045 6.5 24.5V9.16667C6.5 8.45942 6.78095 7.78115 7.28105 7.28105Z"
-          fill="#242424"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M11.68 14.3333C11.68 13.781 12.1277 13.3333 12.68 13.3333H19.3467C19.899 13.3333 20.3467 13.781 20.3467 14.3333C20.3467 14.8855 19.899 15.3333 19.3467 15.3333H12.68C12.1277 15.3333 11.68 14.8855 11.68 14.3333Z"
-          fill="#242424"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M16.0133 10C16.5656 10 17.0133 10.4477 17.0133 11V17.6667C17.0133 18.219 16.5656 18.6667 16.0133 18.6667C15.4611 18.6667 15.0133 18.219 15.0133 17.6667V11C15.0133 10.4477 15.4611 10 16.0133 10Z"
-          fill="#242424"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default PlusBox;
