--- src/assets/SearchIcon.js
@@ -1,29 +1,0 @@
-import React from 'react';
-
-export const SearchIcon = () => {
-  return (
-    <div className="list__icon search">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M11.3333 6.79922C8.82959 6.79922 6.79995 8.82886 6.79995 11.3326C6.79995 13.8362 8.82959 15.8659 11.3333 15.8659C13.837 15.8659 15.8666 13.8362 15.8666 11.3326C15.8666 8.82886 13.837 6.79922 11.3333 6.79922ZM5.19995 11.3326C5.19995 7.94521 7.94594 5.19922 11.3333 5.19922C14.7206 5.19922 17.4666 7.94521 17.4666 11.3326C17.4666 14.7199 14.7206 17.4659 11.3333 17.4659C7.94594 17.4659 5.19995 14.7199 5.19995 11.3326Z"
-          fill="black"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M14.5344 14.5351C14.8468 14.2227 15.3533 14.2227 15.6657 14.5351L18.5657 17.4351C18.8782 17.7475 18.8782 18.254 18.5657 18.5665C18.2533 18.8789 17.7468 18.8789 17.4344 18.5665L14.5344 15.6665C14.2219 15.354 14.2219 14.8475 14.5344 14.5351Z"
-          fill="black"
-        />
-      </svg>
-    </div>
-  );
-};
