--- src/assets/SearchIconDash.js
@@ -1,22 +1,0 @@
-import React from 'react';
-
-export const SearchIconDash = () => {
-  return (
-    <div className="list__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M20.21 18.7892L17.21 15.7892C18.2438 14.4561 18.8033 12.8162 18.8 11.1292C18.8019 9.61138 18.3535 8.12709 17.5115 6.86423C16.6695 5.60138 15.4717 4.61674 14.0698 4.03498C12.6678 3.45322 11.1248 3.30048 9.63607 3.59611C8.14731 3.89174 6.77973 4.62244 5.70646 5.69571C4.63318 6.76898 3.90248 8.13656 3.60685 9.62533C3.31122 11.1141 3.46396 12.6571 4.04572 14.059C4.62749 15.4609 5.61212 16.6587 6.87497 17.5008C8.13783 18.3428 9.62213 18.7912 11.14 18.7892C12.8269 18.7925 14.4668 18.233 15.8 17.1992L18.8 20.1992C18.9883 20.3875 19.2437 20.4933 19.51 20.4933C19.7763 20.4933 20.0317 20.3875 20.22 20.1992C20.4083 20.0109 20.514 19.7555 20.514 19.4892C20.514 19.2229 20.4083 18.9675 20.22 18.7792L20.21 18.7892ZM5.49996 11.1692C5.50185 10.2385 5.73282 9.32258 6.17248 8.50227C6.61214 7.68196 7.24697 6.98248 8.02095 6.46559C8.79492 5.94869 9.68423 5.63027 10.6104 5.53842C11.5366 5.44657 12.4711 5.58413 13.3315 5.93894C14.1919 6.29375 14.9518 6.85491 15.544 7.5729C16.1362 8.29088 16.5426 9.1436 16.7272 10.0558C16.9119 10.968 16.8691 11.9116 16.6028 12.8034C16.3364 13.6952 15.8546 14.5077 15.2 15.1692C14.4061 15.9517 13.3992 16.4828 12.3051 16.6961C11.2111 16.9095 10.0784 16.7956 9.0487 16.3688C8.01899 15.942 7.13793 15.2211 6.51565 14.2963C5.89337 13.3715 5.50748 12.2839 5.49996 11.1692Z"
-          fill="#1B1B1B"
-        />
-      </svg>
-    </div>
-  );
-};
