--- src/assets/Setting.js
@@ -1,27 +1,0 @@
-import React from 'react';
-
-const Settings = () => {
-  return (
-    <div className="nav__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="24" height="24" fill="white" fillOpacity="0.01" />
-        <path
-          d="M18.9799 14.8482C19.0161 14.7746 19.0745 14.7141 19.1468 14.6753C19.2192 14.6365 19.3018 14.6213 19.3832 14.6318C20.0798 14.6318 20.7479 14.3551 21.2405 13.8625C21.7331 13.3699 22.0099 12.7018 22.0099 12.0051C22.0099 11.3085 21.7331 10.6404 21.2405 10.1478C20.7479 9.65521 20.0798 9.37847 19.3832 9.37847H19.2455C19.1906 9.3785 19.1365 9.36612 19.0871 9.34228C19.0377 9.31843 18.9943 9.28372 18.9602 9.24075C18.9497 9.17453 18.9263 9.11102 18.8913 9.05383C18.8615 8.97548 18.8569 8.88981 18.8781 8.80871C18.8992 8.72761 18.9451 8.65511 19.0094 8.6013C19.2729 8.36026 19.4845 8.06802 19.6312 7.74243C19.778 7.41685 19.8568 7.06475 19.8628 6.70768C19.8689 6.3506 19.8021 5.99604 19.6665 5.66567C19.5308 5.33529 19.3293 5.03604 19.0741 4.78621C18.8189 4.53637 18.5154 4.3412 18.1822 4.21262C17.8491 4.08404 17.4932 4.02476 17.1363 4.03839C16.7794 4.05202 16.4291 4.13828 16.1067 4.2919C15.7843 4.44552 15.4966 4.66327 15.2612 4.93184C15.2063 4.98531 15.1362 5.0206 15.0606 5.03292C14.9849 5.04523 14.9073 5.03399 14.8382 5.00071C14.7646 4.96445 14.7041 4.90607 14.6653 4.83374C14.6265 4.76142 14.6113 4.67877 14.6218 4.59736C14.6218 3.90073 14.345 3.23263 13.8524 2.74003C13.3598 2.24744 12.6917 1.9707 11.9951 1.9707C11.2985 1.9707 10.6304 2.24744 10.1378 2.74003C9.64519 3.23263 9.36845 3.90073 9.36845 4.59736V4.73509C9.36848 4.78995 9.3561 4.8441 9.33225 4.89351C9.3084 4.94291 9.27369 4.98628 9.23072 5.02038C9.16573 5.04104 9.10552 5.07449 9.05364 5.11876C8.9753 5.14854 8.88962 5.15317 8.80852 5.13202C8.72742 5.11086 8.65493 5.06496 8.60111 5.00071C8.36738 4.71296 8.07539 4.47794 7.74433 4.31112C7.41327 4.14429 7.05065 4.04943 6.68031 4.03278C6.30997 4.01612 5.94029 4.07805 5.5956 4.21448C5.2509 4.35092 4.93899 4.55877 4.68037 4.82438C4.42176 5.08999 4.22229 5.40733 4.09509 5.75554C3.9679 6.10375 3.91585 6.47495 3.94237 6.84471C3.96889 7.21448 4.07339 7.57444 4.24898 7.90093C4.42458 8.22742 4.66729 8.51305 4.96117 8.73903C5.01765 8.80051 5.05366 8.878 5.06424 8.96082C5.07481 9.04364 5.05942 9.12769 5.0202 9.2014C4.98914 9.2788 4.93428 9.34434 4.86355 9.38854C4.79283 9.43275 4.70988 9.45334 4.62669 9.44734C4.26889 9.42541 3.91039 9.47695 3.57326 9.59879C3.23613 9.72063 2.9275 9.91018 2.66639 10.1558C2.40527 10.4014 2.1972 10.6978 2.05496 11.0269C1.91273 11.3559 1.83936 11.7106 1.83936 12.0691C1.83936 12.4275 1.91273 12.7822 2.05496 13.1113C2.1972 13.4403 2.40527 13.7368 2.66639 13.9824C2.9275 14.228 3.23613 14.4175 3.57326 14.5394C3.91039 14.6612 4.26889 14.7127 4.62669 14.6908H4.76442C4.84484 14.6883 4.92374 14.7132 4.98811 14.7615C5.05248 14.8098 5.0985 14.8785 5.11857 14.9564C5.1526 15.0333 5.16044 15.1192 5.14089 15.201C5.12134 15.2827 5.07548 15.3558 5.01036 15.409C4.74684 15.65 4.53524 15.9422 4.3885 16.2678C4.24176 16.5934 4.16295 16.9455 4.15689 17.3026C4.15084 17.6597 4.21766 18.0142 4.35328 18.3446C4.4889 18.675 4.69046 18.9742 4.94565 19.2241C5.20085 19.4739 5.50431 19.6691 5.83748 19.7976C6.17066 19.9262 6.52656 19.9855 6.88343 19.9719C7.2403 19.9582 7.59064 19.872 7.91304 19.7184C8.23544 19.5647 8.52312 19.347 8.75851 19.0784C8.82191 19.0213 8.90177 18.9858 8.98661 18.977C9.07146 18.9682 9.15693 18.9864 9.23072 19.0292C9.30644 19.0588 9.37045 19.1123 9.41306 19.1815C9.45566 19.2507 9.47454 19.332 9.46683 19.4129C9.4449 19.7707 9.49644 20.1292 9.61828 20.4663C9.74012 20.8035 9.92967 21.1121 10.1753 21.3732C10.4209 21.6343 10.7173 21.8424 11.0464 21.9846C11.3754 22.1269 11.7301 22.2002 12.0886 22.2002C12.447 22.2002 12.8017 22.1269 13.1308 21.9846C13.4598 21.8424 13.7563 21.6343 14.0019 21.3732C14.2475 21.1121 14.437 20.8035 14.5589 20.4663C14.6807 20.1292 14.7322 19.7707 14.7103 19.4129V19.2752C14.7049 19.1933 14.7259 19.1119 14.7701 19.0429C14.8144 18.9738 14.8796 18.9208 14.9563 18.8915C15.0327 18.8654 15.115 18.8617 15.1934 18.8809C15.2719 18.9001 15.3432 18.9413 15.3989 18.9997C15.6401 19.245 15.9281 19.4393 16.2458 19.5711C16.5635 19.7029 16.9045 19.7696 17.2484 19.7671C17.594 19.7709 17.9369 19.7049 18.2564 19.573C18.5759 19.4412 18.8654 19.2462 19.1078 18.9997C19.3555 18.7546 19.5522 18.4627 19.6865 18.141C19.8207 17.8193 19.8898 17.4742 19.8898 17.1256C19.8898 16.7771 19.8207 16.432 19.6865 16.1103C19.5522 15.7886 19.3555 15.4967 19.1078 15.2516C19.0558 15.1978 19.0207 15.1301 19.0067 15.0567C18.9928 14.9833 19.0005 14.9074 19.0291 14.8384L18.9799 14.8482ZM17.1697 14.071C16.979 14.5099 16.9268 14.9965 17.0199 15.4659C17.1131 15.9352 17.3473 16.365 17.6911 16.6977C17.7539 16.7583 17.8038 16.8309 17.8379 16.9111C17.872 16.9914 17.8896 17.0778 17.8896 17.165C17.8896 17.2522 17.872 17.3386 17.8379 17.4188C17.8038 17.4991 17.7539 17.5717 17.6911 17.6323C17.5674 17.7527 17.4014 17.8199 17.2288 17.8192C17.0563 17.8188 16.8907 17.7519 16.7664 17.6323L16.7074 17.5733C16.3761 17.2462 15.9541 17.0263 15.4962 16.9421C15.0383 16.8579 14.5657 16.9133 14.1397 17.1011C13.7209 17.2825 13.364 17.5818 13.1124 17.9627C12.8609 18.3435 12.7256 18.7893 12.7231 19.2457V19.3834C12.7231 19.5565 12.655 19.7227 12.5335 19.846C12.412 19.9694 12.2469 20.0399 12.0738 20.0425C11.9798 20.039 11.8875 20.0162 11.8026 19.9756C11.7177 19.935 11.6421 19.8774 11.5803 19.8064C11.5186 19.7354 11.472 19.6525 11.4435 19.5629C11.415 19.4732 11.4052 19.3786 11.4147 19.285C11.4093 18.8251 11.2657 18.3775 11.0025 18.0002C10.7394 17.623 10.3688 17.3337 9.93904 17.1699C9.64746 17.0419 9.33274 16.9749 9.01429 16.9732C8.37479 16.9802 7.76363 17.2381 7.31238 17.6913C7.25075 17.7539 7.17723 17.8036 7.09614 17.8374C7.01505 17.8712 6.92802 17.8884 6.84017 17.8881C6.66707 17.8855 6.50193 17.8149 6.38044 17.6916C6.25895 17.5682 6.19086 17.4021 6.19088 17.2289C6.19422 17.1299 6.21775 17.0325 6.26004 16.9429C6.30233 16.8532 6.36249 16.7731 6.43682 16.7075C6.76389 16.3762 6.98381 15.9543 7.06801 15.4964C7.15222 15.0385 7.09683 14.5659 6.90903 14.1399C6.72762 13.7211 6.42825 13.3642 6.04743 13.1126C5.66661 12.8611 5.22081 12.7258 4.76442 12.7233H4.62669C4.45357 12.7233 4.28739 12.6552 4.16406 12.5337C4.04073 12.4122 3.97015 12.2471 3.96756 12.074C3.97112 11.98 3.9939 11.8877 4.03451 11.8028C4.07512 11.7179 4.13269 11.6423 4.20367 11.5805C4.27466 11.5187 4.35754 11.4722 4.44722 11.4437C4.53689 11.4152 4.63145 11.4054 4.72507 11.4149C5.18497 11.4095 5.63263 11.2659 6.00984 11.0027C6.38704 10.7396 6.67636 10.369 6.84017 9.93922C7.0325 9.49992 7.08469 9.01199 6.98963 8.54194C6.89456 8.0719 6.65687 7.6426 6.30893 7.31256C6.18239 7.18725 6.10849 7.01834 6.10234 6.84035C6.10493 6.66725 6.17551 6.50212 6.29884 6.38063C6.42217 6.25914 6.58835 6.19105 6.76147 6.19107C6.86054 6.1944 6.95789 6.21794 7.04755 6.26023C7.1372 6.30252 7.21728 6.36267 7.28286 6.43701C7.59056 6.7422 7.97785 6.95475 8.40054 7.05042C8.82322 7.14609 9.26429 7.12103 9.67342 6.97808C9.7545 6.96426 9.83372 6.94115 9.90952 6.90922C10.3298 6.7299 10.6882 6.43103 10.94 6.04974C11.1919 5.66846 11.3262 5.22156 11.3261 4.7646V4.62688C11.3261 4.53864 11.3438 4.45129 11.3781 4.37001C11.4125 4.28873 11.4628 4.21518 11.5261 4.15371C11.5894 4.09224 11.6644 4.04411 11.7467 4.01216C11.8289 3.98022 11.9167 3.96512 12.0049 3.96775C12.0976 3.97008 12.1888 3.99137 12.2729 4.0303C12.3571 4.06923 12.4323 4.12498 12.4941 4.19411C12.5558 4.26324 12.6028 4.34428 12.632 4.43225C12.6613 4.52021 12.6722 4.61324 12.6641 4.70558C12.6639 5.16004 12.7971 5.60457 13.0472 5.98405C13.2972 6.36352 13.6532 6.66126 14.0709 6.84035C14.5107 7.02985 14.9979 7.08052 15.4673 6.98558C15.9367 6.89065 16.3659 6.65463 16.6975 6.30912C16.8237 6.18941 16.9909 6.12267 17.1648 6.12267C17.3387 6.12267 17.506 6.18941 17.6321 6.30912C17.7526 6.43281 17.8197 6.59883 17.819 6.77149C17.8157 6.87056 17.7921 6.96791 17.7499 7.05757C17.7076 7.14723 17.6474 7.2273 17.5731 7.29289C17.2699 7.60184 17.0586 7.98911 16.9631 8.41133C16.8675 8.83354 16.8914 9.27403 17.032 9.68344C17.0458 9.76452 17.0689 9.84375 17.1009 9.91955C17.2802 10.3399 17.5791 10.6982 17.9603 10.9501C18.3416 11.2019 18.7885 11.3362 19.2455 11.3362H19.3832C19.558 11.3362 19.7257 11.4056 19.8493 11.5292C19.9729 11.6528 20.0423 11.8205 20.0423 11.9953C20.04 12.088 20.0187 12.1792 19.9798 12.2633C19.9408 12.3474 19.8851 12.4227 19.816 12.4844C19.7468 12.5462 19.6658 12.5931 19.5778 12.6224C19.4899 12.6516 19.3968 12.6625 19.3045 12.6544C18.8489 12.6551 18.4036 12.7897 18.024 13.0416C17.6444 13.2935 17.3473 13.6515 17.1697 14.071Z"
-          fill="#999999"
-        />
-        <path
-          d="M12.0049 8.56195C11.3239 8.56195 10.6582 8.76389 10.092 9.14223C9.52579 9.52057 9.08446 10.0583 8.82386 10.6875C8.56325 11.3166 8.49506 12.009 8.62792 12.6769C8.76078 13.3448 9.08871 13.9583 9.57025 14.4398C10.0518 14.9214 10.6653 15.2493 11.3332 15.3822C12.0011 15.515 12.6934 15.4468 13.3226 15.1862C13.9518 14.9256 14.4895 14.4843 14.8679 13.9181C15.2462 13.3518 15.4481 12.6861 15.4481 12.0051C15.4481 11.0919 15.0854 10.2162 14.4396 9.57043C13.7939 8.92471 12.9181 8.56195 12.0049 8.56195ZM12.0049 13.4808C11.7131 13.4808 11.4278 13.3942 11.1851 13.2321C10.9425 13.0699 10.7533 12.8395 10.6416 12.5698C10.5299 12.3002 10.5007 12.0035 10.5577 11.7172C10.6146 11.431 10.7551 11.1681 10.9615 10.9617C11.1679 10.7553 11.4308 10.6148 11.7171 10.5578C12.0033 10.5009 12.3 10.5301 12.5697 10.6418C12.8393 10.7535 13.0698 10.9426 13.2319 11.1853C13.3941 11.428 13.4806 11.7133 13.4806 12.0051C13.4806 12.3965 13.3251 12.7718 13.0484 13.0486C12.7717 13.3253 12.3963 13.4808 12.0049 13.4808Z"
-          fill="#999999"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Settings;
