--- src/assets/Silders.js
@@ -1,72 +1,0 @@
-import React from 'react';
-
-const Silders = () => {
-  return (
-    <div className="list__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M18.6667 14.334C19.219 14.334 19.6667 14.7817 19.6667 15.334V19.5007C19.6667 20.0529 19.219 20.5007 18.6667 20.5007C18.1145 20.5007 17.6667 20.0529 17.6667 19.5007V15.334C17.6667 14.7817 18.1145 14.334 18.6667 14.334Z"
-          fill="#242424"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M15.1667 15.334C15.1667 14.7817 15.6145 14.334 16.1667 14.334H21.1667C21.719 14.334 22.1667 14.7817 22.1667 15.334C22.1667 15.8863 21.719 16.334 21.1667 16.334H16.1667C15.6145 16.334 15.1667 15.8863 15.1667 15.334Z"
-          fill="#242424"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M5.33325 12.666C5.88554 12.666 6.33325 13.1137 6.33325 13.666V19.4993C6.33325 20.0516 5.88554 20.4993 5.33325 20.4993C4.78097 20.4993 4.33325 20.0516 4.33325 19.4993V13.666C4.33325 13.1137 4.78097 12.666 5.33325 12.666Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M1.83325 13.666C1.83325 13.1137 2.28097 12.666 2.83325 12.666H7.83325C8.38554 12.666 8.83325 13.1137 8.83325 13.666C8.83325 14.2183 8.38554 14.666 7.83325 14.666H2.83325C2.28097 14.666 1.83325 14.2183 1.83325 13.666Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12 11C12.5523 11 13 11.4477 13 12V19.5C13 20.0523 12.5523 20.5 12 20.5C11.4477 20.5 11 20.0523 11 19.5V12C11 11.4477 11.4477 11 12 11Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M8.5 8.66602C8.5 8.11373 8.94772 7.66602 9.5 7.66602H14.5C15.0523 7.66602 15.5 8.11373 15.5 8.66602C15.5 9.2183 15.0523 9.66602 14.5 9.66602H9.5C8.94772 9.66602 8.5 9.2183 8.5 8.66602Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M18.6667 3.5C19.219 3.5 19.6667 3.94772 19.6667 4.5V12C19.6667 12.5523 19.219 13 18.6667 13C18.1145 13 17.6667 12.5523 17.6667 12V4.5C17.6667 3.94772 18.1145 3.5 18.6667 3.5Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M12 3.5C12.5523 3.5 13 3.94772 13 4.5V8.66667C13 9.21895 12.5523 9.66667 12 9.66667C11.4477 9.66667 11 9.21895 11 8.66667V4.5C11 3.94772 11.4477 3.5 12 3.5Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M5.33325 3.5C5.88554 3.5 6.33325 3.94772 6.33325 4.5V10.3333C6.33325 10.8856 5.88554 11.3333 5.33325 11.3333C4.78097 11.3333 4.33325 10.8856 4.33325 10.3333V4.5C4.33325 3.94772 4.78097 3.5 5.33325 3.5Z"
-          fill="#1B1B1B"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Silders;
