--- src/assets/UserPlus.js
@@ -1,42 +1,0 @@
-import React from 'react';
-
-const Silders = () => {
-  return (
-    <div className="list__icon">
-      <svg
-        width="24"
-        height="24"
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M9.08325 5.5C7.79459 5.5 6.74992 6.54467 6.74992 7.83333C6.74992 9.122 7.79459 10.1667 9.08325 10.1667C10.3719 10.1667 11.4166 9.122 11.4166 7.83333C11.4166 6.54467 10.3719 5.5 9.08325 5.5ZM4.74992 7.83333C4.74992 5.4401 6.69002 3.5 9.08325 3.5C11.4765 3.5 13.4166 5.4401 13.4166 7.83333C13.4166 10.2266 11.4765 12.1667 9.08325 12.1667C6.69002 12.1667 4.74992 10.2266 4.74992 7.83333Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M15.1666 11.1667C15.1666 10.6144 15.6143 10.1667 16.1666 10.1667H21.1666C21.7189 10.1667 22.1666 10.6144 22.1666 11.1667C22.1666 11.7189 21.7189 12.1667 21.1666 12.1667H16.1666C15.6143 12.1667 15.1666 11.7189 15.1666 11.1667Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M18.6666 7.66666C19.2189 7.66666 19.6666 8.11437 19.6666 8.66666V13.6667C19.6666 14.2189 19.2189 14.6667 18.6666 14.6667C18.1143 14.6667 17.6666 14.2189 17.6666 13.6667V8.66666C17.6666 8.11437 18.1143 7.66666 18.6666 7.66666Z"
-          fill="#1B1B1B"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M3.10246 14.7692C3.91511 13.9565 5.01731 13.5 6.16658 13.5H11.9999C13.1492 13.5 14.2514 13.9565 15.064 14.7692C15.8767 15.5819 16.3333 16.6841 16.3333 17.8333V19.5C16.3333 20.0523 15.8855 20.5 15.3333 20.5C14.781 20.5 14.3333 20.0523 14.3333 19.5V17.8333C14.3333 17.2145 14.0874 16.621 13.6498 16.1834C13.2122 15.7458 12.6188 15.5 11.9999 15.5H6.16658C5.54775 15.5 4.95425 15.7458 4.51667 16.1834C4.07908 16.621 3.83325 17.2145 3.83325 17.8333V19.5C3.83325 20.0523 3.38554 20.5 2.83325 20.5C2.28097 20.5 1.83325 20.0523 1.83325 19.5V17.8333C1.83325 16.6841 2.2898 15.5819 3.10246 14.7692Z"
-          fill="#1B1B1B"
-        />
-      </svg>
-    </div>
-  );
-};
-
-export default Silders;
