+++ src/components/ChannelContainer/Channel/ChannelInner.js
@@ -13,7 +13,6 @@ import {
 import { PinnedMessageList } from '../PinnedMessage/PinnedMessageList';
 import { ThreadMessageInput } from '../TeamMessage/ThreadMessageInput';
 import { TeamChannelHeader } from '../TeamMessage/TeamChannelHeader';
-import { Vector } from '../../../assets/Icons';
 
 export const GiphyContext = React.createContext({});
 
@@ -66,13 +65,11 @@ export const ChannelInner = props => {
       <div style={{ display: 'flex', width: '100%' }}>
         <Window>
           <TeamChannelHeader {...{ setIsEditing, setPinsOpen }} />
-
           <MessageList
             disableQuotedMessages
             pinPermissions={pinnedPermissions}
           />
           <MessageInput grow overrideSubmitHandler={overrideSubmitHandler} />
-          <Vector />
         </Window>
         <Thread
           additionalMessageInputProps={{
